Move entrypoint tests into separate exe
Moved the entrypoint tests into a separate executable that we can directly link
to the loader to test with and without using the loader exported entrypoints.

NOTE: This is still run as part of Ctest, it is just a seperate exectuable so
manual testing will need to make sure it triggers the correct executable.