Fix bit field usage for MS remove dev exts from inst ext struct
Fix the bit usage of structs.  MS supports signed structs, but a single bit
of a signed component is the signed bit.  So make it two bits so we can still
set and use the value of 1 without issue.

Also, I had improperly added device extensions which used VkPhysicalDevice
to some of the instance extension structure enables.  This caused issues as
well with improper checks being required.