Implement stdout/stderr redirection for testing
Allows tests to capture stdout/stderr and check if it contains the expected output.

This allows the settings file tests to actually validate the `log_filter` fields.
Before, it used the debug log which doesn't actually validate that the log_filter
was set to the expected value, rather it used the debug_log's filters instead.

This commit adds to InstWrapper and DeviceWrapper an assert to make sure that the
vulkan object hasn't been previously created, which could then get overwritten
in the call to CheckCreate().

This commit also makes the test framework respect
VK_LOADER_DISABLE_DYNAMIC_LIBRARY_UNLOADING during shutdown in order to make
debugging memory leaks slightly easier.