TEMP COMMIT - second shot at removing physical device wrapping in the loader
Wrapping of physical device is still not working. Need a way to map wrapped
and unwrapped handles. I thought I could do the loader_get_physical_device()
on the wrapped handle and have it work, but it appears not. Also physical
device groups fail for some cases, (wrapping isn't handled for sure).