Simplify VkDeviceGroupCreateInfo unwrapping logic
The loader requires unwrapping the VkPhysicalDevices that the application
passes into VkDeviceGroupDeviceCreateInfo even though the pNext chain is
suppose to be const. This is unfortunately undefined behavior with no
known solution that doesn't require different undefind behavior.

The issue with the current implementation was that it would replace a
struct in the pNext chain, requiring quite convoluted pointer casting to
do type-punning, which itself is undefined behavior of a different kind.
By just replacing the pointer to the array of VkPhysicalDevice's, the
loader now doesn't need to type-pune unsafely, and can instead use the
safe memcpy method of iterating the pNext chain, then casting the pNext
pointer to VkDeviceGroupDeviceCreateInfo which is safe to do.